+++ src/main/java/cc/ryanc/halo/service/CategoryService.java
@@ -23,7 +23,7 @@ public interface CategoryService {
      * @param cateId 分类目录编号
      * @return category
      */
-    Category removeByCateId(Integer cateId);
+    Category removeByCateId(Long cateId);
 
     /**
      * 修改分类目录信息
@@ -43,7 +43,7 @@ public interface CategoryService {
      * @param cateId 分类编号
      * @return 返回category实体
      */
-    Category findByCateId(Integer cateId);
+    Category findByCateId(Long cateId);
 
     /**
      * 根据分类目录路径查询，用于验证是否已经存在该路径
